pyproject.toml
[tool.poetry]
name = "library"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.7.1,<4.0"
typer = "^0.3.2"
PyYAML = "^5.4.1"
Jinja2 = "^2.11.2"
requests = "^2.25.1"
boto3 = "^1.16.58"
python-dotenv = "^0.15.0"
GDAL = "^3.2.1"
pandas = "^1.2.1"
beautifulsoup4 = "^4.9.3"
rich = "^9.10.0"
lxml = "^4.6.2"
pydantic = "^1.8.1"
usaddress = "^0.5.10"
xlrd = "^2.0.1"
openpyxl = "^3.0.7"
geopandas = "^0.10.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pre-commit = "^2.9.3"
pdoc3 = "^0.9.2"
SQLAlchemy = "^1.3.22"
psycopg2-binary = "^2.8.6"
black = {version = "^21.7b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
library = "library.cli:run"
